requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
httpx==0.28.1
praw==7.8.1
tweepy==4.14.0
python-dotenv==1.0.1
openai==1.58.1
anthropic==0.40.0
beautifulsoup4==4.12.3
lxml==5.3.0
tenacity==9.0.0
asyncio==3.4.3
playwright==1.48.0
selenium==4.25.0
undetected-chromedriver==3.5.5
fake-useragent==1.5.1
cloudscraper==1.2.71
discord.py==2.4.0
aiohttp==3.10.10
scrapy==2.11.2